requirements.txt
plumbum==1.6.8
rpyc==4.1.2
pycryptodome==3.9.1